+++ gameengine/src/main/java/de/mirkosertic/gameengine/type/EffectCanvas.java
@@ -8,11 +8,11 @@ public interface EffectCanvas {
 
     void drawSingleDot(Position aPosition);
 
-    void fillRect(double aX, double aY, double aWidth, double aHeight);
+    void fillRect(int aX, int aY, int aWidth, int aHeight);
 
-    void fillTriangle(double aX0, double aY0, double aX1, double aY1, double aX2, double aY2);
+    void fillTriangle(int aX0, int aY0, int aX1, int aY1, int aX2, int aY2);
 
-    void fillTriangle(GameResource aTexture, double aX0, double aY0, double aX1, double aY1, double aX2, double aY2, double aU0, double aV0, double aU1, double aV1, double aU2, double aV2);
+    void fillTriangle(GameResource aTexture, int aX0, int aY0, int aX1, int aY1, int aX2, int aY2, int aU0, int aV0, int aU1, int aV1, int aU2, int aV2);
 
-    void drawScaled(GameResource aResource, double aX, double aY, double aWidth, double aHeight);
+    void drawScaled(GameResource aResource, int aX, int aY, int aWidth, int aHeight);
 }
